Normal: major performance improvements
It turns out, we can avoid generating noise for the majority of the chunk column,
because there is a factor subtracted from the noise value, which is computed from the
biome elevation information. If this value is larger than 1, the blocks will always
be stone (below noise-affected territory), and if it's less than 0, it'll always be
air. This allows us to generate 75-90% less noise data per chunk, which is an
enormous performance gain in the most expensive part of the code.

There are additional performance gains from:
- Flood-filling subchunks with stone when below the noise-affected min height
- Filling a biome array once and copying it for every subchunk, instead of individually setting it for every block
- Avoiding double biome picking during biome map calculation

Conservatively, this shows about a 2x performance improvement in PopulationTask timings.
It's probably larger even than that, since PopulationTask's main timings also includes
the timings of population as well as base terrain generation.

This may, however, cause some alterations to the placement of trees,
ores, and other populated objects. It appears the performance
improvements were significant enough to change the order of chunk
generation in many cases, resulting in populators running in a different
order when overlapping chunk boundaries. This is a known issue with
as-yet no known solution. It's described in
https://github.com/pmmp/PocketMine-MP/issues/4649.