Noise: optimise the living F out of getFastNoise3D trilinear lerp
- Avoid divisions inside hot loops, which, it turns out, are actually noticeably slower than multiplying by 1 / number
- Avoid repeated subtractions
- Avoid repeated array lookups - this change alone doubled the performance, but was really difficult to pull off

I have to question the wisdom of the lerp functions being implemented the way they are. I don't think it ever makes
sense to do bilinear or trilinear lerp without large arrays of data present, and as this change shows, there's a ton
of optimisation to be gained by avoiding repeated operations in loops. Similar optimisations are possible to a lesser
extent for getFastNoise1D() and getFastNoise2D().